refactor(frontend): consolidate wallet logic into @/wallet module
Encapsulate Privy auth + wagmi behind a single wallet-kind-agnostic API so
consumers never need to know whether a wallet is external, embedded, or smart.

- Add @/wallet module: useWalletAuth (auth + connect/disconnect + one
  participant address + isLoading), useContractWrite (sponsored vs. self-funded
  writes), useActiveWallet (derives the active wallet and gates gas sponsorship
  on an embedded wallet actually having a smart wallet), plus sponsoredWrite
  helpers and shared types.
- disconnect() tears down both layers: the wagmi connection (external wallets)
  and the Privy session (the only way to drop the embedded wallet).
- Remove hooks/useWalletAuth and hooks/useSponsoredContractWrite; repoint all
  imports to @/wallet.
