+++ frontend/src/wallet/index.ts
@@ -1,0 +1,7 @@
+export { useWalletAuth } from "./useWalletAuth";
+export { useContractWrite } from "./useContractWrite";
+export type {
+  ContractWriteRequest,
+  SponsoredNetworkFeeEstimate,
+  WalletKind,
+} from "./types";
